Crate ruma_serde_macros

Derive Macros

AsRefStr

DeserializeFromCowStr

DisplayAsRefStr

FromString

Outgoing
Derive the Outgoing trait, possibly generating an ‘Incoming’ version of the struct this derive macro is used on. Specifically, if no lifetime variables are used on any of the fields of the struct, this simple implementation will be generated:

SerializeAsRefStr

StringEnum
Shorthand for the derives AsRefStr, FromString, DisplayAsRefStr, SerializeAsRefStr and DeserializeFromCowStr.